requirements.txt
crewai==0.28.0
crewai-tools==0.19.0
streamlit==1.32.0
streamlit-chat==0.1.1
google-generativeai==0.3.1
python-dotenv==1.0.0
pdfplumber==0.10.2
pillow==10.1.0
langchain==0.0.335
python-magic>=0.4.27
python-magic-bin==0.4.14
yt-dlp>=2023.11.14
pytube==15.0.0
embedchain==0.0.68
sentence-transformers==2.2.2
qdrant-client==1.7.0
PyPDF2==3.0.1
watchdog==3.0.0
numpy==1.24.4
opencv-python==4.8.1.78
moviepy==1.0.3
pydub==0.25.1
python-pptx==1.0.1
docx2txt==0.8
typing-extensions==4.7.0
faiss-cpu==1.7.4
tiktoken==0.5.1
llama-index-core==0.10.3
pymupdf==1.23.3
docarray==0.32.1 